test/testapp/requirements.txt
Pillow
django-autocomplete-light==2.0.0a18
django-tagging
sorl-thumbnail==11.12.1b
lxml
